requirements.txt
django-admin
environ
openai
psycopg2